[BAU] refactor extraction
